+++ src/main/java/pages/GameSettingsPage.java
@@ -5,8 +5,6 @@ import game.GameData;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
-import javax.swing.event.ChangeEvent;
-import javax.swing.event.ChangeListener;
 import java.awt.*;
 
 /**
@@ -95,20 +93,20 @@ public class GameSettingsPage extends JFrame {
 		color4CB.setVisible(true);
 		cbPanel.add(color4CB);
 
-		// Use Default Checkbox
-		JCheckBox checkbox = new JCheckBox("Use Default Question Bank Trivia Questions?");
-		checkbox.setSelected(true);
-		checkbox.addChangeListener(new ChangeListener() {
-			public void stateChanged(ChangeEvent e) {
-				JCheckBox source = (JCheckBox) e.getSource();
-				if (source.isSelected()) {
-					GameData.setUseDefaultQuestions(true);
-				} else {
-					GameData.setUseDefaultQuestions(false);
-				}
-			}
-		});
-		cbPanel.add(checkbox);
+//		// Use Default Checkbox
+//		JCheckBox checkbox = new JCheckBox("Use Default Question Bank Trivia Questions?");
+//		checkbox.setSelected(true);
+//		checkbox.addChangeListener(new ChangeListener() {
+//			public void stateChanged(ChangeEvent e) {
+//				JCheckBox source = (JCheckBox) e.getSource();
+//				if (source.isSelected()) {
+//					GameData.setUseDefaultQuestions(true);
+//				} else {
+//					GameData.setUseDefaultQuestions(false);
+//				}
+//			}
+//		});
+//		cbPanel.add(checkbox);
 
 		JPanel buttonPanel = new JPanel();
 		buttonPanel.setBackground(new Color(248, 237, 212));
