+++ src/main/java/pages/QuestionEditorPage.java
@@ -6,6 +6,8 @@ import game.GameData;
 import game.Question;
 
 import javax.swing.*;
+import javax.swing.event.ChangeEvent;
+import javax.swing.event.ChangeListener;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.table.TableCellRenderer;
 import javax.swing.table.TableColumn;
@@ -56,13 +58,26 @@ public class QuestionEditorPage extends JFrame {
         lastUpdatedLabel.setHorizontalAlignment(SwingConstants.RIGHT);
         lastUpdatedLabel.setFont(lastUpdatedLabel.getFont().deriveFont(Font.BOLD));
 
+        // Use Default Checkbox
+        JCheckBox checkbox = new JCheckBox("Use Default Question Bank Trivia Questions?");
+        checkbox.setSelected(true);
+        checkbox.addChangeListener(new ChangeListener() {
+            public void stateChanged(ChangeEvent e) {
+                JCheckBox source = (JCheckBox) e.getSource();
+                if (source.isSelected()) {
+                    GameData.setUseDefaultQuestions(true);
+                } else {
+                    GameData.setUseDefaultQuestions(false);
+                }
+            }
+        });
         // Create a panel to hold the last updated label
         JPanel lastUpdatedPanel = new JPanel(new BorderLayout());
+        lastUpdatedPanel.add(checkbox, BorderLayout.WEST);
         lastUpdatedPanel.add(lastUpdatedLabel, BorderLayout.EAST);
 
         // Question Button Panel
         JPanel questionButtonPanel = new JPanel();
-
         JButton newQuestionButton = new JButton("New Question");
         newQuestionButton.addActionListener(e -> tableModel.addRow(new Object[tableModel.getColumnCount()]));
 
@@ -74,7 +89,6 @@ public class QuestionEditorPage extends JFrame {
 
         JButton deleteQuestionButton = new JButton("Delete Question");
         deleteQuestionButton.addActionListener(e -> deleteSelectedQuestion());
-
         questionButtonPanel.add(newQuestionButton);
         questionButtonPanel.add(saveQuestionButton);
         questionButtonPanel.add(refreshQuestionsButton);
@@ -96,9 +110,9 @@ public class QuestionEditorPage extends JFrame {
         // GUI Control Button Panel
         JPanel buttonPanel = new JPanel();
         JButton nextButton = new JButton("Next");
-        nextButton.addActionListener(e -> controller.showGameplayPage());
+        nextButton.addActionListener(e -> controller.showGameSettingsPage());
         JButton backButton = new JButton("Back");
-        backButton.addActionListener(e -> controller.showPlayerNameInputPage());
+        backButton.addActionListener(e -> controller.showWelcomePage());
         JButton instructionsButton = new JButton("Instructions");
         instructionsButton.addActionListener(e -> controller.showInstructionsPage("EDITOR"));
         buttonPanel.add(instructionsButton);
